Adding trackbar class

+++ recognizer_cpp/src/headers/Recognition.hpp
@@ -66,7 +66,7 @@ public:
 	cv::Mat debugFrameMask;
 	cv::Scalar lowerGreen = cv::Scalar(50, 0, 111);		// 40, 190, 50
     cv::Scalar upperGreen = cv::Scalar(65, 255, 255);	// 65, 255, 255
-	int kernelIterations = 1;
+	double kernelIterations = 1;
 
 	cv::Mat debugFrameCircles;
 	int minRadiusRate = 7;		// 6
